chore: split e.g. PrimitiveImpl off verbs/mod
